Refactor medication tracking to not use modal button
Additionally, I split the modal and view controller code files into their appropriate build targets.
