+++ BridgeApp/BridgeAppTests/CodableTrackedDataTests.swift
@@ -324,12 +324,6 @@ class CodableTrackedDataTests: XCTestCase {
                         { "identifier": "itemC1", "sectionIdentifier" : "c" }
                     ],
             "sections": [{ "identifier": "a" }, { "identifier": "b" }, { "identifier": "c" }],
-            "reminder": {
-                "identifier": "medicationReminders",
-                "type": "medicationReminders",
-                "dataType": "integer",
-                "reminderChoices": [ { "text": "No reminder" } ]
-            }
         }
         """.data(using: .utf8)! // our data in native (JSON) format
